Rename Paramium to Propertium.
