--- src/Paramium/Exceptions/UninitializedDefaultCancellationTokenException.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace Paramium.Exceptions
-{
-    public class UninitializedDefaultCancellationTokenException : Exception
-    {
-        public UninitializedDefaultCancellationTokenException()
-            : base($"The DefaultCancellationToken is uninitialized..")
-        {
-        }
-    }
-}
