--- src/Paramium/WindowsDriver/WindowDriverLocators.cs
@@ -1,9 +1,0 @@
-﻿namespace Paramium.WindowsDriver
-{
-    public static class WindowDriverLocators
-    {
-        public static string AutomationId = "accessibility id";
-        public static string ClassName = "class name";
-        public static string Name = "name";
-    }
-}
