+++ src/Propertium/Exceptions/UninitializedDefaultCancellationTokenException.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Propertium.Exceptions
+{
+    public class UninitializedDefaultCancellationTokenException : Exception
+    {
+        public UninitializedDefaultCancellationTokenException()
+            : base($"The DefaultCancellationToken is uninitialized..")
+        {
+        }
+    }
+}
