+++ src/Propertium/Propertium.csproj
@@ -1,0 +1,29 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>netstandard2.1</TargetFramework>
+        <LangVersion>8</LangVersion>
+        <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+        <Copyright>@Copyright Aurelien Noisette 2019</Copyright>
+        <Description>Get WebElement from SearchProperty.</Description>
+        <PackageProjectUrl>https://github.com/aurnoi1/arium-navigation</PackageProjectUrl>
+        <PackageTags>Arium Test Automation Appium Propertium</PackageTags>
+        <PackageLicenseUrl></PackageLicenseUrl>
+        <Version>1.0.0-beta0001</Version>
+        <Authors>Aurelien Noisette</Authors>
+        <Company>Aurelien Noisette</Company>
+        <PackageLicenseExpression>BSD-3-Clause</PackageLicenseExpression>
+        <PackageReleaseNotes></PackageReleaseNotes>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Appium.WebDriver" Version="4.1.1" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Arium\Arium.csproj" />
+    </ItemGroup>
+
+    <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+    </PropertyGroup>
+</Project>
