+++ src/Propertium/WindowsDriver/WindowDriverLocators.cs
@@ -1,0 +1,9 @@
+﻿namespace Propertium.WindowsDriver
+{
+    public static class WindowDriverLocators
+    {
+        public static string AutomationId = "accessibility id";
+        public static string ClassName = "class name";
+        public static string Name = "name";
+    }
+}
