+++ tests/AUTFacade/AUT.Facade.csproj
@@ -11,7 +11,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Arium\Arium.csproj" />
-    <ProjectReference Include="..\..\src\Paramium\Paramium.csproj" />
+    <ProjectReference Include="..\..\src\Propertium\Propertium.csproj" />
   </ItemGroup>
 
 </Project>
