+++ tests/AUTFacade/POMs/PomYellow.cs
@@ -1,7 +1,7 @@
 ﻿using Arium;
 using Arium.Enums;
-using Paramium;
-using Paramium.WindowsDriver;
+using Propertium;
+using Propertium.WindowsDriver;
 using Arium.Interfaces;
 using OpenQA.Selenium;
 using OpenQA.Selenium.Appium.Interfaces;
