--- tests/ParamiumUnitTests/DataAttributes/InlineAutoMoqDataAttribute.cs
@@ -1,17 +1,0 @@
-﻿using AutoFixture.Xunit2;
-using Xunit;
-using Xunit.Sdk;
-
-namespace Paramium.UnitTests.DataAttributes
-{
-    public class InlineAutoMoqDataAttribute : CompositeDataAttribute
-    {
-        public InlineAutoMoqDataAttribute(params object[] values)
-            : base(
-                  new DataAttribute[] {
-                  new InlineDataAttribute(values),
-                  new AutoMoqDataAttribute() })
-        {
-        }
-    }
-}
